pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hybrid_crypto_project"
version = "0.1.0"
description = "Enterprise-grade Hybrid Cryptography CLI reference implementation"
requires-python = ">=3.11"
authors = [ { name = "hybrid team" } ]
readme = "README.md"
dependencies = [
    "cryptography>=40.0",
]

[project.optional-dependencies]
pq = ["python-oqs>=0.7"]

[tool.setuptools.packages.find]
where = ["."]
include = ["hybrid_crypto_project*"]

[project.urls]
Homepage = "https://example.invalid/"
